Drop requests for unknown hosts
If none of the directives have the default_server parameter then the
first server with the address:port pair will be the default server for
this pair.

To prevent host header attacks or other potential problems, we simply
drop the request (by returning 444) when an unknown server name is used
in a request.